adding more sounds

+++ Lafeef/Model/LevelReportData.swift
@@ -8,7 +8,7 @@ import Foundation
 
 struct LevelReportData: Codable {
     var levelNum: String
-    var collectedMoney: Int
+    var collectedMoney: Float
     var collectedScore: Float
     var isPassed: Bool
 }
